TASK: Extract changing the profile picture to own action
With this change, the code for changing the profile picture is moved to
UserController::changeProfilepictureAction. With that, also the form is
changed to call the new action. Futhermore, the old profile picture is
deleted when it is changed, and updating the user record in the
updateAction no longer has the side-effect of re-uploading the profile
picture even if it's not changed.